// mcp-servers/gatling/src/apiClientGenerated/gatlingEnterpriseSchemas.ts
export type PackageType = "package";
// ...
export type PackageAssets = {
  tests: string[];
};
// ...
export type PackageItemResponse = {
  /**
   * Non-empty trimmed string
   */
  name: string;
  /**
   * Team ID
   */
  teamId: string;
  /**
   * @default package
   */
  _type: PackageType & void;
  /**
   * Package ID
   */
  _id: string;
  _assets: PackageAssets;
  _storage: PackageStorageItem;
};
// ...
export type PackageStorageItem =
  | (Omit<PackageStorageItemManaged, "type"> & {
      type: "managed";
    })
  | (Omit<PackageStorageItemPrivate, "type"> & {
      type: "private";
    });

export type PackageStorageItemManaged = PackageStorageManagedItem & PackageStorageItemMixin;

export type PackageStorageItemMixin = {
  type: string;
};

export type PackageStorageItemPrivate = PackageStoragePrivateItem & PackageStorageItemMixin;